RDM-12049: Apply the security classification filters based on the superset of the <classification>(s)
