+++ src/main/java/uk/gov/hmcts/ccd/domain/service/casedataaccesscontrol/AccessProfileServiceImpl.java
@@ -77,7 +77,7 @@ public class AccessProfileServiceImpl implements AccessProfileService, AccessCon
 
     private Boolean readOnly(RoleAssignment roleAssignment,
                              RoleToAccessProfileDefinition roleToAccessProfileDefinition) {
-        return BooleanUtils.isTrue(roleAssignment.getReadOnly()) ||
-            BooleanUtils.isTrue(roleToAccessProfileDefinition.getReadOnly());
+        return BooleanUtils.isTrue(roleAssignment.getReadOnly())
+            || BooleanUtils.isTrue(roleToAccessProfileDefinition.getReadOnly());
     }
 }
