+++ src/main/java/uk/gov/hmcts/ccd/domain/service/common/AccessControlServiceImpl.java
@@ -13,7 +13,7 @@ import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.ccd.domain.model.aggregated.CaseUpdateViewEvent;
 import uk.gov.hmcts.ccd.domain.model.aggregated.CaseViewActionableEvent;
@@ -30,10 +30,9 @@ import static java.util.Arrays.stream;
 import static java.util.stream.Collectors.toList;
 import static org.springframework.util.CollectionUtils.isEmpty;
 import static uk.gov.hmcts.ccd.domain.model.aggregated.CaseViewField.READONLY;
-import static uk.gov.hmcts.ccd.domain.service.common.AccessControlServiceImpl.QUALIFIER;
 
 @Service
-@Qualifier(QUALIFIER)
+@ConditionalOnProperty(name = "enable-attribute-based-access-control", havingValue = "false")
 public class AccessControlServiceImpl implements AccessControlService {
 
     private static final Logger LOG = LoggerFactory.getLogger(AccessControlServiceImpl.class);
