+++ src/main/java/uk/gov/hmcts/ccd/domain/service/createcase/SubmitCaseTransaction.java
@@ -19,7 +19,7 @@ import uk.gov.hmcts.ccd.domain.model.std.Event;
 import uk.gov.hmcts.ccd.domain.service.AccessControl;
 import uk.gov.hmcts.ccd.domain.service.casedataaccesscontrol.CaseDataAccessControl;
 import uk.gov.hmcts.ccd.domain.service.common.CaseTypeService;
-import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationService;
+import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationServiceImpl;
 import uk.gov.hmcts.ccd.domain.service.common.UIDService;
 import uk.gov.hmcts.ccd.domain.service.message.MessageContext;
 import uk.gov.hmcts.ccd.domain.service.message.MessageService;
@@ -39,7 +39,7 @@ class SubmitCaseTransaction implements AccessControl {
     private final CaseTypeService caseTypeService;
     private final CallbackInvoker callbackInvoker;
     private final UIDService uidService;
-    private final SecurityClassificationService securityClassificationService;
+    private final SecurityClassificationServiceImpl securityClassificationService;
     private final CaseDataAccessControl caseDataAccessControl;
     private final MessageService messageService;
 
@@ -50,7 +50,7 @@ class SubmitCaseTransaction implements AccessControl {
                                  final CaseTypeService caseTypeService,
                                  final CallbackInvoker callbackInvoker,
                                  final UIDService uidService,
-                                 final SecurityClassificationService securityClassificationService,
+                                 final SecurityClassificationServiceImpl securityClassificationService,
                                  final CaseDataAccessControl caseDataAccessControl,
                                  final @Qualifier("caseEventMessageService") MessageService messageService
                                  ) {
