+++ src/main/java/uk/gov/hmcts/ccd/domain/service/createevent/ClassifiedCreateEventOperation.java
@@ -7,17 +7,17 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.ccd.domain.model.definition.CaseDetails;
 import uk.gov.hmcts.ccd.domain.model.std.CaseDataContent;
-import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationService;
+import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationServiceImpl;
 
 @Service
 @Qualifier("classified")
 public class ClassifiedCreateEventOperation implements CreateEventOperation {
     private final CreateEventOperation createEventOperation;
-    private final SecurityClassificationService classificationService;
+    private final SecurityClassificationServiceImpl classificationService;
 
     @Autowired
     public ClassifiedCreateEventOperation(@Qualifier("default") CreateEventOperation createEventOperation,
-                                          SecurityClassificationService classificationService) {
+                                          SecurityClassificationServiceImpl classificationService) {
 
         this.createEventOperation = createEventOperation;
         this.classificationService = classificationService;
