+++ src/main/java/uk/gov/hmcts/ccd/domain/service/getcase/ClassifiedGetCaseOperation.java
@@ -3,9 +3,9 @@ package uk.gov.hmcts.ccd.domain.service.getcase;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.ccd.domain.model.definition.CaseDetails;
-import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationService;
 
 import java.util.Optional;
+import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationServiceImpl;
 
 @Service
 @Qualifier("classified")
@@ -13,10 +13,10 @@ public class ClassifiedGetCaseOperation implements GetCaseOperation {
 
 
     private final GetCaseOperation getCaseOperation;
-    private final SecurityClassificationService classificationService;
+    private final SecurityClassificationServiceImpl classificationService;
 
     public ClassifiedGetCaseOperation(@Qualifier("default") GetCaseOperation getCaseOperation,
-                                      SecurityClassificationService classificationService) {
+                                      SecurityClassificationServiceImpl classificationService) {
         this.getCaseOperation = getCaseOperation;
         this.classificationService = classificationService;
     }
