+++ src/main/java/uk/gov/hmcts/ccd/domain/service/getevents/AuthorisedGetEventsOperation.java
@@ -57,8 +57,8 @@ public class AuthorisedGetEventsOperation implements GetEventsOperation {
     }
 
     @Override
-    public Optional<AuditEvent> getEvent(String jurisdiction, String caseTypeId, Long eventId) {
-        return getEventsOperation.getEvent(jurisdiction, caseTypeId, eventId).flatMap(
+    public Optional<AuditEvent> getEvent(CaseDetails caseDetails, String caseTypeId, Long eventId) {
+        return getEventsOperation.getEvent(caseDetails, caseTypeId, eventId).flatMap(
             event -> secureEvent(caseTypeId, event));
     }
