+++ src/main/java/uk/gov/hmcts/ccd/domain/service/getevents/ClassifiedGetEventsOperation.java
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.ccd.domain.model.definition.CaseDetails;
 import uk.gov.hmcts.ccd.domain.model.std.AuditEvent;
-import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationService;
+import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationServiceImpl;
 import uk.gov.hmcts.ccd.domain.service.getcase.CreatorGetCaseOperation;
 import uk.gov.hmcts.ccd.domain.service.getcase.GetCaseOperation;
 
@@ -18,11 +18,11 @@ import static java.util.Collections.singletonList;
 public class ClassifiedGetEventsOperation implements GetEventsOperation {
 
     private final GetEventsOperation getEventsOperation;
-    private final SecurityClassificationService classificationService;
+    private final SecurityClassificationServiceImpl classificationService;
     private final GetCaseOperation getCaseOperation;
 
     public ClassifiedGetEventsOperation(@Qualifier("default") GetEventsOperation getEventsOperation,
-                                        SecurityClassificationService classificationService,
+                                        SecurityClassificationServiceImpl classificationService,
                                         @Qualifier(CreatorGetCaseOperation.QUALIFIER)
                                         final GetCaseOperation getCaseOperation) {
 
@@ -35,12 +35,12 @@ public class ClassifiedGetEventsOperation implements GetEventsOperation {
     public List<AuditEvent> getEvents(CaseDetails caseDetails) {
         final List<AuditEvent> events = getEventsOperation.getEvents(caseDetails);
 
-        return secureEvents(caseDetails.getJurisdiction(), events);
+        return secureEvents(events, caseDetails.getReferenceAsString());
     }
 
     @Override
     public List<AuditEvent> getEvents(String jurisdiction, String caseTypeId, String caseReference) {
-        return secureEvents(jurisdiction, getEventsOperation.getEvents(jurisdiction, caseTypeId, caseReference));
+        return secureEvents(getEventsOperation.getEvents(jurisdiction, caseTypeId, caseReference), caseReference);
     }
 
     @Override
@@ -49,13 +49,13 @@ public class ClassifiedGetEventsOperation implements GetEventsOperation {
     }
 
     @Override
-    public Optional<AuditEvent> getEvent(String jurisdiction, String caseTypeId, Long eventId) {
-        return getEventsOperation.getEvent(jurisdiction, caseTypeId, eventId).flatMap(
-            event -> secureEvent(jurisdiction, event));
+    public Optional<AuditEvent> getEvent(CaseDetails caseDetails, String caseTypeId, Long eventId) {
+        return getEventsOperation.getEvent(caseDetails, caseTypeId, eventId).flatMap(
+            event -> secureEvent(caseDetails.getReferenceAsString(), event));
     }
 
-    private Optional<AuditEvent> secureEvent(String jurisdiction, AuditEvent event) {
-        return secureEvents(jurisdiction, singletonList(event)).stream().findFirst();
+    private Optional<AuditEvent> secureEvent(String caseReference, AuditEvent event) {
+        return secureEvents(singletonList(event), caseReference).stream().findFirst();
     }
 
     private List<AuditEvent> secureEvents(List<AuditEvent> events, String caseReference) {
@@ -63,14 +63,6 @@ public class ClassifiedGetEventsOperation implements GetEventsOperation {
         if (null == caseDetails) {
             return Lists.newArrayList();
         }
-        return classificationService.applyClassification(caseDetails.getJurisdiction(), events);
-    }
-
-    private List<AuditEvent> secureEvents(String jurisdiction, List<AuditEvent> events) {
-        if (null == events) {
-            return Lists.newArrayList();
-        }
-
-        return classificationService.applyClassification(jurisdiction, events);
+        return classificationService.applyClassification(caseDetails, events);
     }
 }
