+++ src/test/java/uk/gov/hmcts/ccd/domain/service/createevent/ClassifiedCreateEventOperationTest.java
@@ -18,7 +18,7 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import uk.gov.hmcts.ccd.domain.model.definition.CaseDetails;
 import uk.gov.hmcts.ccd.domain.model.std.CaseDataContent;
-import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationService;
+import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationServiceImpl;
 
 class ClassifiedCreateEventOperationTest {
 
@@ -32,7 +32,7 @@ class ClassifiedCreateEventOperationTest {
     private CreateEventOperation createEventOperation;
 
     @Mock
-    private SecurityClassificationService classificationService;
+    private SecurityClassificationServiceImpl classificationService;
 
     private ClassifiedCreateEventOperation classifiedCreateEventOperation;
     private CaseDetails caseDetails;
