+++ src/test/java/uk/gov/hmcts/ccd/domain/service/startevent/ClassifiedStartEventOperationTest.java
@@ -28,7 +28,7 @@ import uk.gov.hmcts.ccd.domain.model.definition.CaseDetails;
 import uk.gov.hmcts.ccd.domain.model.definition.CaseTypeDefinition;
 import uk.gov.hmcts.ccd.domain.model.draft.Draft;
 import uk.gov.hmcts.ccd.domain.service.common.CaseDataService;
-import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationService;
+import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationServiceImpl;
 
 class ClassifiedStartEventOperationTest {
 
@@ -42,7 +42,7 @@ class ClassifiedStartEventOperationTest {
     private StartEventOperation startEventOperation;
 
     @Mock
-    private SecurityClassificationService classificationService;
+    private SecurityClassificationServiceImpl classificationService;
 
     @Mock
     private CaseDefinitionRepository caseDefinitionRepository;
